scripts/arbiter/requirements.txt
PyGithub>=2.1.1
openai>=1.0.0
pyyaml>=6.0.1
tinydb>=4.8.0
python-frontmatter>=1.0.0
requests>=2.31.0
